+++ src/collective/fhirpath/utils.py
@@ -3,7 +3,10 @@ from fhirpath.enums import FHIR_VERSION
 from fhirpath.storage import MemoryStorage
 from plone import api
 from plone.app.fhirfield.interfaces import IFhirResource
+from plone.app.fhirfield.interfaces import IFhirResourceValue
 from plone.behavior.interfaces import IBehavior
+from plone.restapi.services import _no_content_marker
+from pydantic import BaseModel
 from zope.component import getUtility
 from zope.schema import getFields
 
@@ -99,3 +102,39 @@ def find_fhirfield_by_name(fieldname):
     if field_:
         FHIRFIELD_NAMES_MAP.insert(fieldname, field_)
     return field_
+
+
+class FHIRModelServiceMixin:
+    """This is performance optimized plone.restapi service's render mixin class
+    for FHIRModel (fhir.resources) content. Avoiding redundant json serilization
+    and deserialization.
+    User it with your Service Class::
+    >>> from collective.fhirpath.utils import FHIRModelServiceMixin
+    >>> from plone.restapi.services import Service
+    >>> class MyAwesomeService(FHIRModelServiceMixin, Service):
+    >>> .... pass
+    """
+
+    def render(self):
+        """ """
+        self.check_permission()
+        content = self.reply()
+        if content is _no_content_marker:
+            return
+
+        self.request.response.setHeader("Content-Type", self.content_type)
+
+        if IFhirResourceValue.providedBy(content):
+            value = content.foreground_origin()
+            if value:
+                content = value
+            else:
+                return _no_content_marker
+
+        if isinstance(content, BaseModel):
+            return content.json(indent=2, sort_keys=True, separators=(", ", ": "))
+
+        elif isinstance(content, dict):
+            return json.dumps(
+                content, indent=2, sort_keys=True, separators=(", ", ": ")
+            )